testing/web-platform/tests/old-tests/submission/Infraware/Session_History/contents/Session_History/history_replacestate_nooptionalparam.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>history_replaceStateNoOptionalParam</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>
		
		try {
			window.history.replaceState(1, document.title);

			var second;
			second = window.history.state;
			test(function() {
				assert_equals(second, 1, "replaceState method is failed.");
			});
		} catch (e) {
			test(function() {
				assert_unreached("call of replaceState method is failed.");
			});
		}

	</script>
 </body>
</html>
